pyproject.toml
[project]
name = "localstack-pro-samples"
version = "0.1.0"
description = "LocalStack Pro sample applications"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0",
    "pytest-timeout>=2.0",
    "boto3>=1.34",
    "tenacity>=8.0",
    "requests>=2.31",
    "websockets>=12.0",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[tool.pytest.ini_options]
testpaths = ["samples"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--timeout=300",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Markers for filtering tests
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
